chore: init logger
