+++ cmd/root.go
@@ -1,9 +1,7 @@
 package cmd
 
 import (
-	"fmt"
 	"log"
-	"os"
 	"time"
 
 	"github.com/FarmerChillax/ALiCloudDDNS/client"
@@ -12,23 +10,19 @@ import (
 	"github.com/spf13/cobra"
 )
 
+const VERSION = "0.2.0"
+
 var rootCmd = &cobra.Command{
 	Use:   "fddns",
 	Short: "fddns 是一个 ddns 客户端",
 	Long: `fddns 是一个轻量 ddns 客户端
 目前仅支持阿里云服务，后续有需要会做更多的云服务商支持`,
-	Version: flog.VERSION,
+	Version: VERSION,
 	// 如果有相关的 action 要执行，请取消下面这行代码的注释
 	Run: func(cmd *cobra.Command, args []string) {
 		if logFileName != "" {
 			flog.SetLoggerWithFile(logFileName)
 		}
-		var stop string
-		go func() {
-			log.Println("[Start] 程序运行中, 按任意键关闭...")
-			fmt.Scanln(&stop)
-			os.Exit(0)
-		}()
 
 		ddnsConfig := config.Get()
 		ddnsClient := client.New(ddnsConfig)
