--- log/version.go
@@ -1,3 +1,0 @@
-package log
-
-const VERSION = "0.2.0"
