+++ proto/ddns_service.proto
@@ -5,7 +5,7 @@ package ddns_server;
 option go_package = "./ddns_server";
 
 service DdnsServer {
-    rpc HeartBeatServer (HeartBeatClient) returns (HeartBeat);
+    rpc HeartBeatServer (stream HeartBeatClient) returns (stream HeartBeat);
 }
 
 message HeartBeatClient {
