+++ server/server.go
@@ -1,0 +1,8 @@
+package server
+
+type DdnsServer struct {
+}
+
+func NewDdnsServer() *DdnsServer {
+	return &DdnsServer{}
+}
